Upgrade cc65 version
Took me *forever* to figure out how to disable executable headers in the
binary output.

Turns out we needed `-Wl -D__EXEHDR__=0` as an argument to cl65. The -Wl
means pass the following flag to the linker, and the other flag gets
interpreted by the linker when it's reading the .cfg file (details are a
little unclear to me; I don't fully understand the cfg file -- but hey,
it works!)